Create a data frame with all possible transition and associated next state for each scenario and each state for one particular week, used in Bellman. For each next state, Bellman value is calculated.

Create a data frame with all possible transition and associated next state for each scenario and each state for one particular week, used in Bellman. For each next state, Bellman value is calculated.

Usage

build_all_possible_decisions(
  Data_week,
  decision_space,
  f_next_value,
  mcyears,
  lvl_high,
  lvl_low,
  E_max,
  P_max,
  next_week_values,
  niveau_max,
  overflow_cost
)

Arguments

Data_week

Data frame generated in Grid_Matrix code containing list of states and years evaluated (we suppose there is only one week at a time) as weel as inflows

decision_space

Simulation constraints values for the week generated by get_Reward or runWaterValuesSimulation

f_next_value

List of approximated Bellman functions for the next week, generated by get_bellman_values_interpolation

mcyears

Vector of monte carlo years to take into account

lvl_high

Double, top rule curve for the end of the week in MWh

lvl_low

Double, bottom rule curve for the end of the week in MWh

E_max

Double, weekly maximum turbining power for the week

P_max

Double, weekly maximum pumping power for the week

next_week_values

Vector of Bellman values for the next week

niveau_max

Reservoir capacity in MWh

overflow_cost

Cost for overflow (equal to spillage cost of the area)

Value

Data frame with same format as Data week with 3 more columns : next_state, control (transition to next state) and next value (Bellman value of next state)