Calculate and plot watervalues with value_node_gen. Penalties taken into account. Used in Grid_Matrix

Calculate and plot watervalues with value_node_gen. Penalties taken into account. Used in Grid_Matrix

Usage

build_data_watervalues(
  watervalues,
  statesdt,
  reservoir,
  penalty_level_high,
  penalty_level_low,
  force_final_level = FALSE,
  penalty_final_level_low = 0,
  penalty_final_level_high = 0,
  final_level = NULL,
  plot_watervalues = T
)

Arguments

watervalues

Data frame generated in Grid_Matrix

statesdt

Data frame of possible states, generated in Grid_Matrix

reservoir

Data frame describing rule curves, generated by readReservoirLevels

penalty_level_high

Penalty for violating the bottom rule curve

penalty_level_low

Penalty for violating the top rule curve

force_final_level

Binary. Whether final level should be constrained

penalty_final_level_low

Penalties for both bottom rule curve to constrain final level

penalty_final_level_high

Penalties for top rule curve to constrain final level

final_level

Double. Final level to consider

plot_watervalues

Binary. Plot watervalues or not ?

Value

Data frame with water value (vu) for each week (weeks) and each state (states). vu_pen corresponds to water value without penalties