Compute reward function with the 5 simulations method. Called for each area of MultiStock_H2_Investment_reward_compute_once. If they are several areas, the trajectories of the other storage are fixed on their optimal trend.

Compute reward function with the 5 simulations method. Called for each area of MultiStock_H2_Investment_reward_compute_once. If they are several areas, the trajectories of the other storage are fixed on their optimal trend.

Usage

calculateRewardsSimulations(
  area,
  list_areas,
  list_pumping,
  list_efficiency,
  opts,
  mcyears,
  path_to_antares,
  prefix,
  launch_sims = T,
  sim_number = 5,
  optimal_traj,
  inflow,
  candidates_types
)

Arguments

area

Character. Name of the area where the reward is computed

list_areas

Vector of characters of the names of areas to optimize.

list_pumping

Vector of boolean. True if pumping for the area with corresponding index in list_areas.

list_efficiency

Vector of numeric with pumping efficiency for the area with corresponding index in list_areas.

opts

List of study parameters returned by the function antaresRead::setSimulationPath(simulation="input") in input mode.

mcyears

Vector of integers. Monte Carlo years to run simulations

path_to_antares

Character containing the Antares Solver path, argument passed to runSimulation.

prefix

Character. Prefix of the simulation.

launch_sims

Boolean. True to launch simulations, false if simulations already run.

sim_number

Integer. Number of simulations.

optimal_traj

Data frame containing optimal trajectory for all areas

inflow

Generated by get_inflow()

candidates_types

Data_frame with column names : c(index, name, type, TOTEX, Marg_price, Part_fixe, Prix_fixe, Borne_min, Borne_max, Points_nb, Zone).

Value

a data_frame containing the rewards returned by the function antaresWaterValues::get_Reward()