Generate control values

Generate a list of control values, eg difference between final and initial level for each week for the given area. Used to run simulations in runWaterValuesSimulation() and also to estimate reward functions in functions such as Grid_Matrix() and get_Reward().

Usage

constraint_generator(
  area,
  nb_disc_stock,
  pumping = T,
  efficiency = NULL,
  opts,
  max_hydro = NULL,
  inflow = NULL,
  mcyears = NULL
)

Arguments

area

Character. The Antares area concerned by water values computation.

nb_disc_stock

Integer. Number of constraint values wanted for each week.

pumping

Boolean. True to take into account the pumping capacity.

efficiency

Double between 0 and 1. Pumping efficiency ratio. Get it with getPumpEfficiency().

opts

List of study parameters returned by the function antaresRead::setSimulationPath(simulation="input") in input mode.

max_hydro

Weekly pumping and turbining maximum powers, generated by get_max_hydro().

inflow

Weekly inflow, generated by get_inflow().

mcyears

Vector of integer. Monte Carlo years used to compute water values.

Value

A dplyr::tibble() with 3 columns : "week" (integer), "u" (double), "sim" (character). For each week from 1 to 52, nb_disc_stock different controls values are generated from maximum pumping power to maximum generating power. The controls are equally spaced between maximum pumping power (negative value) and 0 and between 0 and maximum generating power (positive value). Controls values are stored in column "u". Column "sim" indexes control values in the form "u_i" with i the index.