Generate constraint values for a week at a time used in constraint_generator depending on the maximum powers for the week

Generate constraint values for a week at a time used in constraint_generator depending on the maximum powers for the week

Usage

constraint_week(pumping, efficiency, nb_disc_stock, res_cap, hydro, week)

Arguments

pumping

Boolean. True to take into account the pumping.

efficiency

between 0 and 1. the pumping efficiency ratio.

nb_disc_stock

Number of constraint values wanted for each week

res_cap

Double, reservoir capacity

hydro

Pumping and turbining maximum powers for the week, generated by get_max_hydro

week

Current week for which to compute controls

Value

List of constraint values for the week