Compute reward functions based on one given simulation

Reward functions are computed for all weeks from 1 to 52 and for all scenarios in mcyears for the given area. For a specific week and a specific scenario, the reward function is evaluated for all controls listed in possible_controls based on marginal prices and on the real control done in the simulation. This function is mainly used in get_Reward().

Usage

get_local_reward(
  simu,
  possible_controls,
  max_hydro_hourly,
  area,
  mcyears,
  u0,
  efficiency
)

Arguments

simu

List of simulation parameters returned by the function antaresRead::setSimulationPath() with the simulation selected.

possible_controls

If method_old=FALSE, controls for which to compute reward, generated by constraint_generator().

max_hydro_hourly

Maximum hourly pumping and generating power generated by the function get_max_hydro() with timeStep="hourly".

area

Character. The Antares area concerned by water values computation.

mcyears

Vector of integer. Monte Carlo years used to compute water values.

u0

A dplyr::tibble() with columns "week", "u" and "mcYear" (optional) that gives constraint values per week (and per scenario) used in the simulation. Could be extracted from the simulation_values output of runWaterValuesSimulation() by filtering column "sim" to the corresponding simulation.

efficiency

Double between 0 and 1. Pumping efficiency ratio. Get it with getPumpEfficiency().

Value

A dplyr::tibble() with columns "week", "mcYear", "u" and "reward". Controls in column "u" correspond to possible_controls.