Take into account overall cost in local reward

Modify local reward to take into account overall cost of the simulation, mainly used in get_Reward().

Usage

reward_offset(simu, df_reward, u0 = c(), mcyears, expansion = F)

Arguments

simu

List of simulation parameters returned by the function antaresRead::setSimulationPath() with the simulation selected.

df_reward

A dplyr::tibble() computed by the function get_local_reward().

u0

A dplyr::tibble() with columns "week", "u" and "mcYear" (optional) that gives constraint values per week (and per scenario) used in the simulation. Could be extracted from the simulation_values output of runWaterValuesSimulation() by filtering column "sim" to the corresponding simulation.

mcyears

Vector of integer. Monte Carlo years used to compute water values.

expansion

Binary. True if mode expansion (ie linear relaxation) of Antares is used to run simulations, argument passed to runSimulation. It is recommended to use mode expansion, it will be faster (only one iteration is done) and results will be smoother as the cost result will correspond to the linear relaxation of the problem.

Value

A dplyr::tibble() with columns "week", "mcYear", "u" and "reward".