Run Antares simulations in order to compute water values for multiple areas

Each simulation evaluates for each week and each scenario the cost of a set of controls for all areas in list_areas.

Usage

runWaterValuesSimulationMultiStock(
  list_areas,
  list_pumping,
  list_efficiency,
  simulation_name = "wv_sim_%s",
  mcyears = NULL,
  path_solver = NULL,
  show_output_on_console = FALSE,
  overwrite = FALSE,
  opts,
  file_name = NULL,
  launch_simulations = NULL,
  constraint_values = NULL,
  expansion = TRUE
)

Arguments

list_areas

Vector of areas concerned by simulations.

list_pumping

Named vector of binary to tell if pumping is available in areas.

list_efficiency

Named vector of pumping efficiency.

simulation_name

Character. The name of the simulation, s is a placeholder for the constraint value defined by nb_disc_stock.

mcyears

Vector of integer. Monte Carlo years used to compute water values.

path_solver

Character containing the Antares Solver path, argument passed to runSimulation.

show_output_on_console

Argument passed to runSimulation.

overwrite

If area or cluster already exists, should they be overwritten?

opts

List of study parameters returned by the function antaresRead::setSimulationPath(simulation="input") in input mode.

file_name

the Rdata file name.

launch_simulations

Boolean or vector of booleans with the same length as nb_disc_stock. True to to run the simulations.

constraint_values

Constraint values for each area to use to run simulations, generated by the function constraint_generator with an additional column "area".

expansion

Binary. True if mode expansion (ie linear relaxation) of Antares is used to run simulations, argument passed to runSimulation. It is recommended to use mode expansion, it will be faster (only one iteration is done) and results will be smoother as the cost result will correspond to the linear relaxation of the problem.

Value

simulation_names

Vector of character. Names of simulations.

simulation_values

A dplyr::tibble() generated by constraint_generator() describing controls for each simulation.

list_areas

Parameter list_areas.

mcyears

Parameter mcyears.

list_pumping

Parameter list_pumping.

list_efficiency

Parameter list_efficiency.

expansion

Parameter expansion.

See also