Setup a simulation before running it for calculating Water Values, used in runWaterValuesSimulation

Setup a simulation before running it for calculating Water Values, used in runWaterValuesSimulation

Usage

setupWaterValuesSimulation(
  area,
  fictive_area_name = paste0("watervalue_", area),
  thermal_cluster = "water_value_cluster",
  overwrite = FALSE,
  remove_areas = NULL,
  opts = antaresRead::simOptions(),
  reset_hydro = T,
  link_from = NULL,
  pumping = F,
  max_load = 1e+08,
  ...
)

Arguments

area

The area concerned by the simulation.

fictive_area_name

Name of the fictive area to create.

thermal_cluster

Name of the thermal cluster to create.

overwrite

If area or cluster already exists, overwrite them ?

remove_areas

Character vector of area(s) to remove from the created district.

opts

List of simulation parameters returned by the function antaresRead::setSimulationPath

reset_hydro

Boolean. True to reset hydro inflow to 0 before the simulation.

link_from

area that will be linked to the created fictive area. If it's NULL it will takes the area concerned by the simulation.

pumping

Boolean. True to take into account the pumping.

max_load

the maximum load to put it in the fictive areas.

...

further arguments passed to or from other methods.

Value

The result of antaresRead::simOptions