Computes the total reward for a specific candidate following the reward function

Computes the total reward for a specific candidate following the reward function

Usage

total_cost_loop(
  area,
  mc_years,
  candidate_pool,
  candidates_types,
  storage_vol,
  df_reward,
  cvar,
  opts,
  penalty_low,
  penalty_high,
  penalty_final_level,
  storage_annual_cost,
  final_level,
  compute_optimal_traj = F,
  max_hydro_weekly,
  inflow
)

Arguments

area

Character. Name of the area where the total reward is computed

mc_years

Vector of integers. Monte Carlo years to compute reward

candidate_pool

Vector of integers. Capacity of the cluster candidates

candidates_types

Data_frame with column names : c("index", "name", "type", "TOTEX", "Marg_price"). It is a parameter of MultiStock_H2_Investment_reward_compute_once.

storage_vol

Integer. Volume of the storage candidate

df_reward

List. Rewards and decision space from the output of the function antaresWaterValues::get_Reward().

cvar

Numeric in [0,1]. The probability used in cvar algorithm.

opts

List of study parameters returned by the function antaresRead::setSimulationPath(simulation="input") in input mode.

penalty_low

Integer. Penalty for lower guide curve.

penalty_high

Integer. Penalty for higher guide curve.

penalty_final_level

Integer. Penalty for higher and lower final level.

storage_annual_cost

Numeric. Total annual cost of the storage candidate in eur/MWh

final_level

Numeric in [0, 100]. Final and initial level in percentage of H2 storage.

compute_optimal_traj

Boolean.

max_hydro_weekly

Data.frame. Generated by get_max_hydro()

inflow

Generated by get_inflow()

Value

The total reward for the specified candidates in euros. Because of the Bellman values it is usually negative and large